BorrowInfo modified
m_borrowed and m_reserved has been deleted from BorrowInfo class
