+++ Library-Mangement-System/Library-Mangement-System.vcxproj.filters
@@ -18,12 +18,12 @@
     <ClCompile Include="main.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="BorrowedBook.cpp">
+    <ClCompile Include="BorrowedInfo.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="BorrowedBook.h">
+    <ClInclude Include="BorrowedInfo.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
